% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneids.R
\name{annotategene}
\alias{annotategene}
\title{annotate gene
This function converts a EntrezID to a Gene name}
\usage{
annotategene(x)
}
\description{
annotate gene
This function converts a EntrezID to a Gene name
}
\examples{
annotategene()
annotategene_function()
}
\keyword{EntrezID}
\keyword{Gene}
