% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{annotationTable}
\alias{annotationTable}
\title{Metadata for RNA-seq time course}
\format{
An object of class \code{data.frame} with 286 rows and 7 columns.
}
\usage{
annotationTable
}
\description{
.
}
\examples{
annotationTable
}
\keyword{datasets}
