% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneids.R
\name{any2entrez}
\alias{any2entrez}
\title{Any to Entrez
This function converts a anyhting to Entrez ID}
\usage{
any2entrez(x)
}
\description{
Any to Entrez
This function converts a anyhting to Entrez ID
}
\examples{
any2entrez('TFF1')
any2entrez_function()

}
\keyword{EntrezID}
\keyword{Gene}
