% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneids.R
\name{ens2eg}
\alias{ens2eg}
\title{Ensemble 2 Entrez
This function converts a Ensemble ID to Entrez ID}
\usage{
ens2eg(x)
}
\description{
Ensemble 2 Entrez
This function converts a Ensemble ID to Entrez ID
}
\examples{
ens2eg('ENSG00000160182')
ens2eg_function()
}
\keyword{Convert}
\keyword{Ensemble}
\keyword{Entrez}
