% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{expmat_brca_a}
\alias{expmat_brca_a}
\title{First half of the TCGA Breast Cancer Expression Matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10265 rows and 1154 columns.
}
\usage{
expmat_brca_a
}
\description{
.
}
\examples{
expmat_brca_a
}
\keyword{datasets}
