% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{expmat_brca_b}
\alias{expmat_brca_b}
\title{Second half of the TCGA Breast Cancer Expression Matrix}
\format{
An object of class \code{matrix} with 10266 rows and 1154 columns.
}
\usage{
expmat_brca_b
}
\description{
.
}
\examples{
expmat_brca_b
}
\keyword{datasets}
