% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{expmat_metabric}
\alias{expmat_metabric}
\title{METABRIC Breast Cancer Expression Matrix}
\format{
An object of class \code{matrix} with 24420 rows and 996 columns.
}
\usage{
expmat_metabric
}
\description{
.
}
\examples{
expmat_metabric
}
\keyword{datasets}
