% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1b.R
\name{figure1b}
\alias{figure1b}
\title{Figure 1b}
\usage{
figure1b()
}
\description{
This function allows you generate figure1b - qPLEX-RIME results.
}
\examples{
figure1b()
}
\keyword{ER}
\keyword{estrogen}
\keyword{qPlexRIME}
