% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1d.R
\name{figure1d}
\alias{figure1d}
\title{Figure 1d}
\usage{
figure1d()
}
\description{
This function allows you generate figure1d
}
\examples{
figure1d()
}
\keyword{ER}
\keyword{estrogen}
\keyword{luciferase}
\keyword{siZMIZ1}
