% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure1e.R
\name{figure1e}
\alias{figure1e}
\title{Figure 1e
This function allows you generate figure1e}
\usage{
figure1e()
}
\description{
Figure 1e
This function allows you generate figure1e
}
\examples{
figure1e()
}
\keyword{ER}
\keyword{VIPER}
\keyword{ZMIZ1}
