% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2a.R
\name{figure2a}
\alias{figure2a}
\title{Figure 2a}
\usage{
figure2a()
}
\description{
This function allows you generate figure2a
}
\examples{
figure2a()
}
\keyword{ER}
\keyword{estrogen}
\keyword{luciferase}
\keyword{siZMIZ1}
