% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2a.R
\name{figure2a_PLA}
\alias{figure2a_PLA}
\title{Figure 2a}
\usage{
figure2a_PLA()
}
\description{
This function allows you generate figure2a - PLA results.
}
\examples{
figure2a_PLA()
}
\keyword{Analysis}
\keyword{PLA}
