% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2c.R
\name{figure2b}
\alias{figure2b}
\title{Figure 2c}
\usage{
figure2b()
}
\description{
This function allows you generate figure2c - luciferase activity assay
}
\examples{
figure2b()
}
\keyword{ER}
\keyword{estrogen}
\keyword{luciferase}
\keyword{siZMIZ1}
