% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2d.R
\name{figure2c}
\alias{figure2c}
\title{Figure 2d}
\usage{
figure2c()
}
\description{
This function allows you generate figure2d - siZMIZ1 cell growth
}
\examples{
figuredc()
}
\keyword{cell}
\keyword{growth}
\keyword{siZMIZ1}
