% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure3a.R
\name{figure3a}
\alias{figure3a}
\title{Figure 3a
This function allows you generate figure1e}
\usage{
figure3a()
}
\description{
Figure 3a
This function allows you generate figure1e
}
\examples{
figure3a()
}
\keyword{ER}
\keyword{VIPER}
\keyword{ZMIZ1}
