% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure3b.R
\name{figure3b}
\alias{figure3b}
\title{Figure 3b
This function allows you generate figure3b}
\usage{
figure3b()
}
\description{
Figure 3b
This function allows you generate figure3b
}
\examples{
figure3b()
}
\keyword{ER}
\keyword{VIPER}
\keyword{ZMIZ1}
