% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure3c.R
\name{figure3c_ChIPqPCR}
\alias{figure3c_ChIPqPCR}
\title{Figure 3c}
\usage{
figure3c_ChIPqPCR()
}
\description{
This function allows you generate figure3c - ZMIZ1 ChIP-qpcr at E2F2 promoter
}
\examples{
figure3c_ChIPqPCR()
}
\keyword{ChIP}
\keyword{siZMIZ1}
