% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure3e.R
\name{figure3e_E2F2Timecourse}
\alias{figure3e_E2F2Timecourse}
\title{Figure 3e}
\usage{
figure3e_E2F2Timecourse()
}
\description{
This function allows you generate figure3e - E2F2 expression by timepoint
}
\examples{
figure3e_E2F2Timecourse()
}
\keyword{E2F2}
\keyword{RNA}
\keyword{siZMIZ1}
\keyword{timecourse}
