% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure4.R
\name{figure4_cellGrowth}
\alias{figure4_cellGrowth}
\title{Figure 4}
\usage{
figure4_cellGrowth()
}
\description{
This function allows you generate figure2d - siZMIZ1 cell growth
}
\examples{
figure4_cellFrowth()
}
\keyword{cell}
\keyword{growth}
\keyword{siZMIZ1}
