% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure4c.R
\name{figure4c}
\alias{figure4c}
\title{Figure 4c
This function allows you generate figure4 - ER and ZMIZ1 activity}
\usage{
figure4c()
}
\description{
Figure 4c
This function allows you generate figure4 - ER and ZMIZ1 activity
}
\examples{
figure4c()
}
\keyword{ER}
\keyword{VIPER}
\keyword{ZMIZ1}
