% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure4e.R
\name{figure4e}
\alias{figure4e}
\title{Figure 4e
This function allows you generate figure4e - ZMIZ1 and KI67}
\usage{
figure4e()
}
\description{
Figure 4e
This function allows you generate figure4e - ZMIZ1 and KI67
}
\examples{
figure4e()
}
\keyword{ER}
\keyword{VIPER}
\keyword{ZMIZ1}
