% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure5c.R
\name{figure5c_VIPER}
\alias{figure5c_VIPER}
\title{Figure 5c
This function allows you generate figure4 - ER and ZMIZ1 activity}
\usage{
figure5c_VIPER()
}
\description{
Figure 5c
This function allows you generate figure4 - ER and ZMIZ1 activity
}
\examples{
figure4c()
}
\keyword{ER}
\keyword{VIPER}
\keyword{ZMIZ1}
