% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure5e.R
\name{figure5e_ki67}
\alias{figure5e_ki67}
\title{Figure 5e
This function allows you generate figure4e - ZMIZ1 and KI67}
\usage{
figure5e_ki67()
}
\description{
Figure 5e
This function allows you generate figure4e - ZMIZ1 and KI67
}
\examples{
figure4e()
}
\keyword{ER}
\keyword{VIPER}
\keyword{ZMIZ1}
