% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure2c.R
\name{figure_removed_2c}
\alias{figure_removed_2c}
\title{Figure 2c}
\usage{
figure_removed_2c()
}
\description{
This function allows you generate the old figure2c - luciferase activity assay
}
\examples{
figure_removed_2c
}
\keyword{ER}
\keyword{estrogen}
\keyword{luciferase}
\keyword{siZMIZ1}
