% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS1.R
\name{figures1}
\alias{figures1}
\title{Figure S1}
\usage{
figures1()
}
\description{
This function allows you generate figure s1. P-value will vary as seed is
set.
}
\examples{
figures1()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
