% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS11.R
\name{figureS11}
\alias{figureS11}
\title{Figure S11}
\usage{
figureS11()
}
\description{
This function allows you generate figure S11. GSEA results Part 1. P-value will vary as seed is
not set.
}
\examples{
figureS11()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
