% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS12.R
\name{figureS12_E2_gene_set_volcano}
\alias{figureS12_E2_gene_set_volcano}
\title{Figure S12}
\usage{
figureS12_E2_gene_set_volcano()
}
\description{
This function allows you generate figure S12 - E2 response of gene sets.
}
\examples{
figureS12_E2_gene_set_volcano()
}
\keyword{GSEA}
\keyword{RNAseq}
\keyword{ZMIZ1}
