% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS13.R
\name{figureS13mcf7}
\alias{figureS13mcf7}
\title{Figure S13mcf7}
\usage{
figureS13mcf7()
}
\description{
This function allows you generate figure S13. GSEA results Part 2. P-value will vary as seed is
not set.
}
\examples{
figureS13mcf7()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
