% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS13.R
\name{figureS13t47d}
\alias{figureS13t47d}
\title{Figure S13t47d}
\usage{
figureS13t47d()
}
\description{
This function allows you generate figure S3. GSEA results Part 2. P-value will vary as seed is
not set.
}
\examples{
figureS13mcf7()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
