% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS14.R
\name{figureS14_Survival_METABRIC}
\alias{figureS14_Survival_METABRIC}
\title{Figure S14_Survival_METABRIC}
\usage{
figureS14_Survival_METABRIC()
}
\description{
This function allows you generate figure S14_Survival_METABRIC.
}
\examples{
figureS14_Survival_METABRIC()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
