% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS14.R
\name{figureS14_Survival_TCGA}
\alias{figureS14_Survival_TCGA}
\title{Figure S14_Survival_TCGA}
\usage{
figureS14_Survival_TCGA()
}
\description{
This function allows you generate figure S14_Survival_TCGA.
}
\examples{
figureS14_Survival_TCGA()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
