% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS15.R
\name{figureS15_METABRIC}
\alias{figureS15_METABRIC}
\title{Figure S15}
\usage{
figureS15_METABRIC()
}
\description{
This function allows you generate figure S15 metabric - luminal vs basal ZMIZ1 expression
}
\examples{
figureS15_METABRIC()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
