% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS15.R
\name{figureS15_TCGA}
\alias{figureS15_TCGA}
\title{Figure S15 TCGA}
\usage{
figureS15_TCGA()
}
\description{
This function allows you generate figure S15 TCGA - luminal vs basal ZMIZ1 expression
}
\examples{
figureS15_TCGA()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
