% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS3.R
\name{figureS3_METABRIC}
\alias{figureS3_METABRIC}
\title{Figure S3_METABRIC}
\usage{
figureS3_METABRIC()
}
\description{
This function allows you generate figure s3_METABRIC.
}
\examples{
figureS3_METABRIC()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
