% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS3.R
\name{figureS3_TCGA}
\alias{figureS3_TCGA}
\title{Figure S3_TCGA}
\usage{
figureS3_TCGA()
}
\description{
This function allows you generate figure s3_TCGA.
}
\examples{
figureS3_TCGA()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
