% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS4.R
\name{figureS4}
\alias{figureS4}
\title{Figure S4}
\usage{
figureS4()
}
\description{
This function allows you generate figure s4. P-value will vary as seed is
not set.
}
\examples{
figureS4()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
