% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS5.R
\name{figureS5_METABRIC}
\alias{figureS5_METABRIC}
\title{Figure S5_METABRIC}
\usage{
figureS5_METABRIC()
}
\description{
This function allows you generate figure S5_METABRIC.
}
\examples{
figureS5_METABRIC()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
