% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS5.R
\name{figureS5_TCGA}
\alias{figureS5_TCGA}
\title{Figure S5_TCGA}
\usage{
figureS5_TCGA()
}
\description{
This function allows you generate figure S5_TCGA.
}
\examples{
figureS5_TCGA()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
