% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS6.R
\name{figureS6_METABRIC}
\alias{figureS6_METABRIC}
\title{Figure S6}
\usage{
figureS6_METABRIC()
}
\description{
This function allows you generate figure S6 metabric - luminal vs basal ZMIZ1 expression
}
\examples{
figureS6_METABRIC()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
