% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS6.R
\name{figureS6_TCGA}
\alias{figureS6_TCGA}
\title{Figure S6 TCGA}
\usage{
figureS6_TCGA()
}
\description{
This function allows you generate figure S6 TCGA - luminal vs basal ZMIZ1 expression
}
\examples{
figureS6_TCGA()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
