% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS7.R
\name{figureS7}
\alias{figureS7}
\title{Figure S7}
\usage{
figureS7()
}
\description{
This function allows you generate figure S7. GSEA results Part 1. P-value will vary as seed is
not set.
}
\examples{
figureS4()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
