% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS8.R
\name{figureS8mcf7}
\alias{figureS8mcf7}
\title{Figure S8mcf7}
\usage{
figureS8mcf7()
}
\description{
This function allows you generate figure S8. GSEA results Part 2. P-value will vary as seed is
not set.
}
\examples{
figureS8mcf7()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
