% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS1.R
\name{figureS9_siZMIZ1_ESR1_ZMIZ1_reads}
\alias{figureS9_siZMIZ1_ESR1_ZMIZ1_reads}
\title{Figure S9}
\usage{
figureS9_siZMIZ1_ESR1_ZMIZ1_reads()
}
\description{
This function allows you generate figure S9.
}
\examples{
figureS9_siZMIZ1_ESR1_ZMIZ1_reads()
}
\keyword{RNAseq}
\keyword{ZMIZ1}
