% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{proteinData}
\alias{proteinData}
\title{MSigDB gene sets.}
\format{
An object of class \code{data.frame} with 418 rows and 3 columns.
}
\usage{
proteinData
}
\description{
.
}
\examples{
proteinData
}
\keyword{datasets}
