% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rawcounts}
\alias{rawcounts}
\title{RNA-seq time course data}
\format{
An object of class \code{matrix} with 28395 rows and 286 columns.
}
\usage{
rawcounts
}
\description{
.
}
\examples{
rawcounts
}
\keyword{datasets}
