% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{subtypes_brca}
\alias{subtypes_brca}
\title{TCGA samples subtype metadata}
\format{
An object of class \code{character} of length 1154.
}
\usage{
subtypes_brca
}
\description{
.
}
\examples{
subtypes_brca
}
\keyword{datasets}
