% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{subtypes_metabric}
\alias{subtypes_metabric}
\title{METABRIC samples subtype metadata}
\format{
An object of class \code{character} of length 996.
}
\usage{
subtypes_metabric
}
\description{
.
}
\examples{
subtypes_metabric
}
\keyword{datasets}
