% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneids.R
\name{sym2eg}
\alias{sym2eg}
\title{Symbol 2 Entrez
This function converts a gene symbol to Entrez ID}
\usage{
sym2eg(ids)
}
\arguments{
\item{ids}{is vector of EntrezIDs}
}
\description{
Symbol 2 Entrez
This function converts a gene symbol to Entrez ID
}
\examples{
sym2eg('TFF1')
}
\keyword{EntrezID}
\keyword{Gene}
\keyword{Symbol}
