% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vipermat_brca}
\alias{vipermat_brca}
\title{VIPER Activity Matrix for TCGA BRCA patient samples.}
\format{
An object of class \code{matrix} with 4386 rows and 1037 columns.
}
\usage{
vipermat_brca
}
\description{
.
}
\examples{
vipermat_brca
}
\keyword{datasets}
