% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vipermat_metabric}
\alias{vipermat_metabric}
\title{VIPER Activity Matrix for METABRIC patient samples.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 3375 rows and 996 columns.
}
\usage{
vipermat_metabric
}
\description{
.
}
\examples{
vipermat_metabric
}
\keyword{datasets}
