% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneids.R
\name{any2entrez}
\alias{any2entrez}
\title{Any to Entrez
This function converts a anything to Entrez ID}
\usage{
any2entrez(x)
}
\arguments{
\item{x}{is type of ID that could be converted to an EntrezID}
}
\description{
Any to Entrez
This function converts a anything to Entrez ID
}
\examples{
any2entrez('TFF1')

}
\keyword{EntrezID}
\keyword{Gene}
