% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneids.R
\name{eg2ens}
\alias{eg2ens}
\title{Entrez 2 Ensemble
This function converts a Entrez ID to Ensemble ID}
\usage{
eg2ens(x)
}
\description{
Entrez 2 Ensemble
This function converts a Entrez ID to Ensemble ID
}
\examples{
ens2eg('7031')
ens2eg_function()
}
\keyword{Convert}
\keyword{Ensemble}
\keyword{Entrez}
