% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneids.R
\name{eg2refseq}
\alias{eg2refseq}
\title{Entrez 2 RegSeq
This function converts a Entrez ID to RefSeq}
\usage{
eg2refseq(x)
}
\arguments{
\item{x}{is vector of EntrezIDs}
}
\description{
Entrez 2 RegSeq
This function converts a Entrez ID to RefSeq
}
\examples{
eg2refseq('7031')


}
\keyword{Convert}
\keyword{Entrez}
\keyword{Gene}
\keyword{Symbol}
