% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneids.R
\name{eg2sym}
\alias{eg2sym}
\title{Entrez 2 Symbol
This function converts a Entrez ID to Gene Symbol}
\usage{
eg2sym(x)
}
\description{
Entrez 2 Symbol
This function converts a Entrez ID to Gene Symbol
}
\examples{
eg2sym('7031')
eg2sym_function()

}
\keyword{Convert}
\keyword{Entrez}
\keyword{Gene}
\keyword{Symbol}
