% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figureS4.R
\name{figureS4}
\alias{figureS4}
\title{Figure S4}
\usage{
figureS4()
}
\description{
This function allows you generate figure s4 - E2 response of gene sets.
}
\examples{
figureS4()
}
\keyword{GSEA}
\keyword{RNAseq}
\keyword{ZMIZ1}
